+++ templates/sitemap.xml.php
@@ -1,30 +1,18 @@
 <?php namespace ProcessWire; ?>
-<?php /** @var $pages \ProcessWire\PageArray */  ?>
-<?php /** @var $languages \ProcessWire\PageArray */  ?>
+<?php /** @var $items \SeoMaestro\SitemapItem[] */  ?>
 <?php echo "<?xml version=\"1.0\" encoding=\"utf-8\"?>\n"; ?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml">
-<?php foreach ($pages as $page): ?>
-<?php if ($hasLanguageSupport && $hasLanguageSupportPageNames): ?>
-<?php foreach ($languages ?: [] as $language): ?>
-<?php if (!$page->viewable($language)) continue; ?>
+<?php foreach ($items as $item): ?>
     <url>
-        <loc><?= $baseUrl ? $baseUrl . $page->localUrl($language) : $page->localHttpUrl($language) ?></loc>
-        <lastmod><?= date('c', $page->modified) ?></lastmod>
-        <changefreq><?= $page->seoMaestroSitemapData->changeFrequency ?></changefreq>
-        <priority><?= $page->seoMaestroSitemapData->priority ?></priority>
-<?php foreach ($languages as $language): ?>
-<?php if (!$page->viewable($language)) continue; ?>
-        <xhtml:link rel="alternate" hreflang="<?= $language->isDefault() ? $defaultLanguageCode : $language->name ?>" href="<?= $baseUrl ? $baseUrl . $page->localUrl($language) : $page->localHttpUrl($language) ?>"/>
-<?php endforeach ?>
-    </url>
-<?php endforeach ?>
-<?php else: ?>
-    <url>
-        <loc><?= $baseUrl ? $baseUrl . $page->url : $page->httpUrl ?></loc>
-        <lastmod><?= date('c', $page->modified) ?></lastmod>
-        <changefreq><?= $page->seoMaestroSitemapData->changeFrequency ?></changefreq>
-        <priority><?= $page->seoMaestroSitemapData->priority ?></priority>
-    </url>
+        <loc><?= $item->loc ?></loc>
+<?php if ($item->lastmod): ?>
+        <lastmod><?= $item->lastmod ?></lastmod>
 <?php endif ?>
+        <changefreq><?= $item->changefreq ?></changefreq>
+        <priority><?= $item->priority ?></priority>
+<?php foreach ($item->alternates as $langCode => $url): ?>
+        <xhtml:link rel="alternate" hreflang="<?= $langCode ?>" href="<?= $url ?>"/>
+<?php endforeach ?>
+    </url>
 <?php endforeach ?>
 </urlset>